+++ apps/ungdomsytelse-deltaker/env.schema.ts
@@ -1,15 +1,37 @@
 import { commonEnvSchema, ungDeltakelseOpplyserEnvSchema } from '@navikt/sif-common-env';
 import * as z from 'zod';
 
+/**
+ * For at disse skal bli tilgjengelige i appen, må disse filene oppdateres:
+ * - appEnv.ts - på server
+ * - getAppSettings.cjs (lokalt)
+ *
+ * I tillegg til at env filene må inneholde verdiene
+ */
+
 export enum AppEnvKey {
     'SIF_PUBLIC_PERSONALIA_URL' = 'SIF_PUBLIC_PERSONALIA_URL',
     'VELG_SCENARIO' = 'VELG_SCENARIO',
+    'SIF_PUBLIC_URL_RETT_OG_PLIKT' = 'SIF_PUBLIC_URL_RETT_OG_PLIKT',
+    'SIF_PUBLIC_URL_PERSONOPPLYSNINGER' = 'SIF_PUBLIC_URL_PERSONOPPLYSNINGER',
+    'SIF_PUBLIC_URL_PERSONVERN' = 'SIF_PUBLIC_URL_PERSONVERN',
+    'SIF_PUBLIC_URL_OM_UNGDOMSPROGRAMYTELSEN' = 'SIF_PUBLIC_URL_OM_UNGDOMSPROGRAMYTELSEN',
+    'SIF_PUBLIC_URL_SKATTEETATEN' = 'SIF_PUBLIC_URL_SKATTEETATEN',
+    'SIF_PUBLIC_MINSIDE_URL' = 'SIF_PUBLIC_MINSIDE_URL',
+    'SIF_PUBLIC_URL_ENDRE_KONTONUMMER' = 'SIF_PUBLIC_URL_ENDRE_KONTONUMMER',
 }
 
 export const appEnvSchema = z
     .object({
         [AppEnvKey.SIF_PUBLIC_PERSONALIA_URL]: z.string().min(1),
         [AppEnvKey.VELG_SCENARIO]: z.string().optional(),
+        [AppEnvKey.SIF_PUBLIC_URL_RETT_OG_PLIKT]: z.string(),
+        [AppEnvKey.SIF_PUBLIC_URL_PERSONOPPLYSNINGER]: z.string(),
+        [AppEnvKey.SIF_PUBLIC_URL_PERSONVERN]: z.string(),
+        [AppEnvKey.SIF_PUBLIC_URL_OM_UNGDOMSPROGRAMYTELSEN]: z.string(),
+        [AppEnvKey.SIF_PUBLIC_URL_SKATTEETATEN]: z.string(),
+        [AppEnvKey.SIF_PUBLIC_MINSIDE_URL]: z.string(),
+        [AppEnvKey.SIF_PUBLIC_URL_ENDRE_KONTONUMMER]: z.string(),
     })
     .merge(commonEnvSchema)
     .merge(ungDeltakelseOpplyserEnvSchema);