+++ apps/ungdomsytelse-deltaker/src/apps/søknad/SøknadRouter.tsx
@@ -3,10 +3,9 @@ import { Navigate, Route, Routes, useLocation } from 'react-router-dom';
 import { usePrevious } from '@navikt/sif-common-hooks';
 import { useSøknadContext } from './hooks/context/useSøknadContext';
 import KvitteringPage from './pages/KvitteringPage';
-import VelkommenSteg from './pages/VelkommenPage';
+import VelkommenPage from './pages/VelkommenPage';
 import BarnSteg from './steg/barn/BarnSteg';
 import KontonummerSteg from './steg/kontonummer/KontonummerSteg';
-import OppstartSteg from './steg/oppstart/OppstartSteg';
 import OppsummeringSteg from './steg/oppsummering/OppsummeringSteg';
 import { Steg } from './types';
 import { getStegRoute, SøknadRoutes } from './utils/routeUtils';
@@ -24,7 +23,7 @@ const SøknadRouter = () => {
      */
     const determineRedirectPath = (): string | null => {
         if (pathname === SøknadRoutes.VELKOMMEN && søknadStartet && previousSøknadStartet === true) {
-            return getStegRoute(Steg.OPPSTART);
+            return getStegRoute(Steg.KONTONUMMER);
         }
         if (pathname !== SøknadRoutes.KVITTERING && søknadSendt && previousSøknadSendt === true) {
             return SøknadRoutes.KVITTERING;
@@ -38,35 +37,24 @@ const SøknadRouter = () => {
         return null;
     };
 
-    const initialRedirectPath = determineRedirectPath();
-    /**
-     * Å sette initialRedirectPath som default value på state gir bare redirectPath===null første
-     * gang komponenten rendres. Derfor både logikk på initialRedirectPath og redirectPath.
-     * */
-
-    const [redirectPath, setRedirectPath] = useState<string | null>();
+    const [redirectPath, setRedirectPath] = useState<string | null>(determineRedirectPath());
 
     useEffect(() => {
         const newRedirectPath = determineRedirectPath();
         if (newRedirectPath !== redirectPath) {
             setRedirectPath(newRedirectPath);
         }
-    }, [redirectPath, pathname, søknadStartet, previousSøknadStartet, previousSøknadSendt]);
+    }, [pathname, søknadStartet, previousSøknadStartet, previousSøknadSendt]);
 
     if (redirectPath && redirectPath !== pathname) {
         return <Navigate to={redirectPath} replace={true} />;
     }
 
-    if (initialRedirectPath !== null && initialRedirectPath !== pathname) {
-        return <Navigate to={initialRedirectPath} replace={true} />;
-    }
-
     return (
         <Routes>
-            <Route index={true} element={<VelkommenSteg />} />
-            <Route path="soknad" element={<Navigate to="/soknad/barn" replace={true} />} />
+            <Route index={true} element={<VelkommenPage />} />
+            <Route path="soknad" element={<Navigate to="/soknad/kontonummer" replace={true} />} />
             <Route path="soknad/barn" element={<BarnSteg />} />
-            <Route path="soknad/oppstart" element={<OppstartSteg />} />
             <Route path="soknad/kontonummer" element={<KontonummerSteg />} />
             <Route path="soknad/oppsummering" element={<OppsummeringSteg />} />
             <Route path="soknad/kvittering" element={<KvitteringPage />} />