+++ apps/ungdomsytelse-deltaker/src/apps/søknad/context/SøknadContext.tsx
@@ -1,7 +1,7 @@
 import React, { createContext, useState } from 'react';
 import { RegistrertBarn, Søker } from '@navikt/sif-common-api';
 import { YesOrNo } from '@navikt/sif-common-core-ds/src';
-import { DeltakelsePeriode } from '@navikt/ung-common';
+import { DeltakelsePeriode, formaterKontonummer } from '@navikt/ung-common';
 import { MellomlagringDTO } from '../api/mellomlagring/mellomlagring';
 import { useSøknadNavigation } from '../hooks/utils/useSøknadNavigation';
 import { Spørsmål, Steg, SøknadContextType, SøknadSvar } from '../types';
@@ -21,7 +21,6 @@ const initialData: SøknadSvar = {};
 // const initialData: SøknadSvar = {
 //     barn: YesOrNo.YES,
 //     kontonummer: YesOrNo.YES,
-//     oppstart: YesOrNo.YES,
 //     harForståttRettigheterOgPlikter: true,
 // };
 
@@ -54,7 +53,7 @@ export const SøknadProvider = ({
             harForståttRettigheterOgPlikter,
         });
         setSøknadStartet(true);
-        gotoSteg(Steg.OPPSTART);
+        gotoSteg(Steg.KONTONUMMER);
     };
 
     return (
@@ -62,7 +61,15 @@ export const SøknadProvider = ({
             value={{
                 svar,
                 søknadSendt,
-                kontonummer,
+                kontonummerInfo: kontonummer
+                    ? {
+                          harKontonummer: true,
+                          kontonummerFraRegister: kontonummer,
+                          formatertKontonummer: formaterKontonummer(kontonummer),
+                      }
+                    : {
+                          harKontonummer: false,
+                      },
                 barn,
                 søknadStartet,
                 deltakelsePeriode,