+++ apps/ungdomsytelse-deltaker/src/apps/søknad/pages/VelkommenPage.tsx
@@ -1,4 +1,4 @@
-import { BodyLong, Box, Button, Checkbox, CheckboxGroup, Link, VStack } from '@navikt/ds-react';
+import { BodyLong, Box, Button, Checkbox, CheckboxGroup, VStack } from '@navikt/ds-react';
 import { useState } from 'react';
 import { ArrowRightIcon } from '@navikt/aksel-icons';
 import Page from '@navikt/sif-common-core-ds/src/components/page/Page';
@@ -6,8 +6,9 @@ import SøknadHeader from '../components/søknad-header/SøknadHeader';
 import VelkommenMelding from '../components/VelkommenMelding';
 import { useSøknadContext } from '../hooks/context/useSøknadContext';
 import { Spørsmål } from '../types';
+import ExternalLink from '../components/external-link/ExternalLink';
 
-const VelkommenSteg = () => {
+const VelkommenPage = () => {
     const { søker, deltakelsePeriode, startSøknad, svar } = useSøknadContext();
 
     const [infoStemmer, setInfoStemmer] = useState<boolean>(svar[Spørsmål.FORSTÅR_PLIKTER] || false);
@@ -26,30 +27,33 @@ const VelkommenSteg = () => {
     return (
         <Page title="Velkommen - Søknad om ungdomsprogramytelse">
             <VStack gap="8">
-                <SøknadHeader tittel="Søknad om ungdomsprogramytelse" />
+                <SøknadHeader />
 
                 <VelkommenMelding fornavn={søker.fornavn} startdato={deltakelsePeriode.programPeriode.from} />
 
                 <div>
-                    <BodyLong>
-                        Det er viktig at du gir oss riktige opplysninger slik at vi kan behandle saken din.{' '}
-                        <Link href="https://www.nav.no/endringer">
-                            Les mer om viktigheten av å gi riktige opplysninger.
-                        </Link>
-                    </BodyLong>
-                    <Box paddingBlock="4 8">
-                        <CheckboxGroup error={error} name="bekreftelse" legend="Bekreftelse" hideLegend={true}>
-                            <Checkbox
-                                value="bekrefter"
-                                onChange={(evt) => {
-                                    setError(undefined);
-                                    setInfoStemmer(evt.target.checked);
-                                }}>
-                                Jeg bekrefter at jeg vil svare så riktig som jeg kan
-                            </Checkbox>
-                        </CheckboxGroup>
-                    </Box>
-                    <form onSubmit={(evt) => handleSubmit(evt)}>
+                    <form onSubmit={handleSubmit}>
+                        <BodyLong>
+                            Det er viktig at du gir oss riktige opplysninger slik at vi kan behandle saken din.{' '}
+                            <ExternalLink href="https://www.nav.no/endringer">
+                                Les mer om viktigheten av å gi riktige opplysninger
+                            </ExternalLink>
+                            .
+                        </BodyLong>
+
+                        <Box paddingBlock="4 8">
+                            <CheckboxGroup error={error} name="bekreftelse" legend="Bekreftelse" hideLegend={true}>
+                                <Checkbox
+                                    value="bekrefter"
+                                    onChange={(evt) => {
+                                        setError(undefined);
+                                        setInfoStemmer(evt.target.checked);
+                                    }}>
+                                    Jeg vil svare så godt jeg kan på spørsmålene i søknaden.
+                                </Checkbox>
+                            </CheckboxGroup>
+                        </Box>
+
                         <Button variant="primary" icon={<ArrowRightIcon aria-hidden />} iconPosition="right">
                             Start søknad
                         </Button>
@@ -60,4 +64,4 @@ const VelkommenSteg = () => {
     );
 };
 
-export default VelkommenSteg;
+export default VelkommenPage;