+++ apps/ungdomsytelse-deltaker/src/apps/søknad/steg/kontonummer/KontonummerSteg.tsx
@@ -1,8 +1,10 @@
-import { Alert, BodyLong, Heading, Radio, RadioGroup, VStack } from '@navikt/ds-react';
+import { Alert, BodyLong, BodyShort, GuidePanel, Heading, Link, Radio, RadioGroup, VStack } from '@navikt/ds-react';
 import { useState } from 'react';
 import { YesOrNo } from '@navikt/sif-common-core-ds/src';
 import { getYesOrNoValidator } from '@navikt/sif-validation';
 import AriaLiveRegion from '../../../../components/aria-live-region/AriaLiveRegion';
+import getLenker from '../../../../utils/lenker';
+import ExternalLink from '../../components/external-link/ExternalLink';
 import SkjemaFooter from '../../components/steg-skjema/SkjemaFooter';
 import SøknadSteg from '../../components/søknad-steg/SøknadSteg';
 import { useSøknadContext } from '../../hooks/context/useSøknadContext';
@@ -10,16 +12,14 @@ import { useSøknadNavigation } from '../../hooks/utils/useSøknadNavigation';
 import { Spørsmål, Steg } from '../../types';
 
 const KontonummerSteg = () => {
-    const { setSpørsmålSvar, svar, kontonummer } = useSøknadContext();
+    const { setSpørsmålSvar, svar, kontonummerInfo } = useSøknadContext();
     const { gotoSteg } = useSøknadNavigation();
 
     const infoStemmer = svar[Spørsmål.KONTONUMMER];
     const [error, setError] = useState<string | undefined>(undefined);
 
-    const harKontonummer = kontonummer !== undefined;
-
     const handleOnSubmit = () => {
-        if (harKontonummer) {
+        if (kontonummerInfo.harKontonummer) {
             const err = getYesOrNoValidator()(infoStemmer);
             if (err) {
                 setError('Du må svare på spørsmålet');
@@ -27,7 +27,7 @@ const KontonummerSteg = () => {
             }
             setError(undefined);
         }
-        gotoSteg(Steg.OPPSUMMERING);
+        gotoSteg(Steg.BARN);
     };
 
     return (
@@ -38,12 +38,15 @@ const KontonummerSteg = () => {
                     handleOnSubmit();
                 }}>
                 <VStack gap="8">
-                    {harKontonummer ? (
+                    <GuidePanel>
+                        For å få pengene inn på bankkontoen din, må du ha registrert kontonummeret ditt hos Nav.
+                    </GuidePanel>
+                    {kontonummerInfo.harKontonummer ? (
                         <>
                             <VStack gap="4">
                                 <RadioGroup
                                     name="barnOk"
-                                    legend={`Er kontonummeret ditt ${kontonummer}?`}
+                                    legend={`Er kontonummeret ditt ${kontonummerInfo.formatertKontonummer}?`}
                                     error={error}
                                     value={infoStemmer}
                                     onChange={(value: YesOrNo) => {
@@ -59,13 +62,17 @@ const KontonummerSteg = () => {
                                 </RadioGroup>
                                 <AriaLiveRegion visible={infoStemmer === YesOrNo.NO}>
                                     <Alert variant="info">
-                                        <BodyLong spacing>
-                                            Gå til personopplysninger på min side for å endre kontonummer.
-                                        </BodyLong>
-                                        <BodyLong>
-                                            Vi anbefaler å endre kontonummer før du sender inn søknaden for å unngå
-                                            utbetaling til feil konto.
-                                        </BodyLong>
+                                        <BodyShort spacing>
+                                            Gå til{' '}
+                                            <ExternalLink href={getLenker().personopplysninger}>
+                                                personopplysninger på Min side
+                                            </ExternalLink>{' '}
+                                            for å endre bankkontonummeret ditt.
+                                        </BodyShort>
+                                        <BodyShort>
+                                            Vi anbefaler at du endrer kontonummeret ditt før du sender inn søknaden,
+                                            slik at pengene ikke blir forsinket.
+                                        </BodyShort>
                                     </Alert>
                                 </AriaLiveRegion>
                             </VStack>
@@ -77,20 +84,19 @@ const KontonummerSteg = () => {
                                     Du har ikke registrert kontonummer hos oss
                                 </Heading>
                                 <BodyLong spacing>
-                                    For å få utbetalt penger til rett konto må du registrere kontonummer hos oss. Gå til
-                                    personopplysninger på min side for å legge inn kontonummeret ditt.
+                                    Registrer bankkontonummeret ditt hos Nav slik at du får pengene utbetalt til rett
+                                    konto. Gå til{' '}
+                                    <Link href={getLenker().endreKontonummer}>personopplysninger på Min side</Link> for
+                                    å legge inn kontonummeret ditt.
                                 </BodyLong>
                                 <BodyLong>
-                                    Du kan fremdeles sende inn søknaden, men vi anbefaler å legge inn kontonummer med en
-                                    gang for å unngå forsinkelser i utbetalingen.
+                                    Du kan fremdeles sende inn søknaden, men vi anbefaler at du legger inn kontonummeret
+                                    med én gang slik at pengene ikke blir forsinket.
                                 </BodyLong>
                             </Alert>
                         </>
                     )}
-                    <SkjemaFooter
-                        forrige={{ tittel: 'Forrige steg', onClick: () => gotoSteg(Steg.BARN) }}
-                        submit={{ tittel: 'Neste steg', erSendInn: false }}
-                    />
+                    <SkjemaFooter submit={{ tittel: 'Neste steg', erSendInn: false }} />
                 </VStack>
             </form>
         </SøknadSteg>