+++ apps/ungdomsytelse-deltaker/src/apps/søknad/utils/stegUtils.ts
@@ -1,10 +1,6 @@
-import { Spørsmål, Steg, SøknadSvar } from '../types';
+import { KontonummerInfo, Spørsmål, Steg, SøknadSvar } from '../types';
 
-export const søknadSteg = [Steg.OPPSTART, Steg.BARN, Steg.KONTONUMMER, Steg.OPPSUMMERING];
-
-export const getStegIndex = (steg: Steg): number => {
-    return søknadSteg.findIndex((s) => s === steg);
-};
+export const søknadSteg = [Steg.KONTONUMMER, Steg.BARN, Steg.OPPSUMMERING];
 
 export const getSkjemaStegIndex = (steg: Steg): number => {
     return søknadSteg.findIndex((s) => s === steg);
@@ -18,40 +14,26 @@ export const getStegFraPath = (path: string): Steg | undefined => {
     return;
 };
 
-export const getNesteSkjemaSteg = (steg: Steg): Steg | undefined => {
-    const index = getSkjemaStegIndex(steg);
-    return index > steg.length - 1 ? søknadSteg[index + 1] : undefined;
-};
-
-export const getForrigeSkjemaSteg = (steg: Steg): Steg | undefined => {
-    const index = getSkjemaStegIndex(steg);
-    return index > 0 ? søknadSteg[index - 1] : undefined;
-};
-
-export const getTilgjengeligeSteg = (svar: SøknadSvar, harKontonummer: boolean): Steg[] => {
+export const getTilgjengeligeSteg = (svar: SøknadSvar, kontonummerInfo: KontonummerInfo): Steg[] => {
     const tilgjengeligeSteg: Steg[] = [];
 
     const velkommenOk: boolean = svar[Spørsmål.FORSTÅR_PLIKTER] === true;
-    const oppstartOk: boolean = velkommenOk && svar[Spørsmål.OPPSTART] !== undefined;
-    const barnOk: boolean = oppstartOk && svar[Spørsmål.BARN] !== undefined;
-    const kontonummerOk: boolean = barnOk && (harKontonummer ? svar[Spørsmål.KONTONUMMER] !== undefined : true);
+    const kontonummerOk: boolean = kontonummerInfo.harKontonummer ? svar[Spørsmål.KONTONUMMER] !== undefined : true;
+    const barnOk: boolean = kontonummerOk && svar[Spørsmål.BARN] !== undefined;
 
     if (velkommenOk) {
-        tilgjengeligeSteg.push(Steg.OPPSTART);
-    }
-    if (oppstartOk) {
-        tilgjengeligeSteg.push(Steg.BARN);
-    }
-    if (barnOk) {
         tilgjengeligeSteg.push(Steg.KONTONUMMER);
     }
     if (kontonummerOk) {
+        tilgjengeligeSteg.push(Steg.BARN);
+    }
+    if (barnOk) {
         tilgjengeligeSteg.push(Steg.OPPSUMMERING);
     }
 
     return tilgjengeligeSteg;
 };
 
-export const erStegTilgjengelig = (steg: Steg, svar: SøknadSvar, harKontonummer: boolean): boolean => {
-    return getTilgjengeligeSteg(svar, harKontonummer).find((s) => s === steg) !== undefined;
+export const erStegTilgjengelig = (steg: Steg, svar: SøknadSvar, kontonummerInfo: KontonummerInfo): boolean => {
+    return getTilgjengeligeSteg(svar, kontonummerInfo).find((s) => s === steg) !== undefined;
 };