+++ apps/ungdomsytelse-deltaker/src/main.tsx
@@ -1,29 +1,12 @@
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
-import { getMaybeEnv, getRequiredEnv } from '@navikt/sif-common-env';
+import { enableMocking } from '../mock/msw/mocking';
 import App from './App';
-
-async function enableMocking() {
-    const ENV = getMaybeEnv('ENV');
-    if (ENV !== 'development') {
-        return;
-    }
-    const { worker } = await import('../mock/msw/browser');
-    if (document.location.pathname === '/') {
-        document.location.replace(getRequiredEnv('PUBLIC_PATH'));
-    }
-    return worker.start();
-}
-
-const queryClient = new QueryClient();
 
 enableMocking().then(() =>
     createRoot(document.getElementById('root')!).render(
         <StrictMode>
-            <QueryClientProvider client={queryClient}>
-                <App />
-            </QueryClientProvider>
+            <App />
         </StrictMode>,
     ),
 );