+++ packages/ung-deltakelse-opplyser-api/src/client/zod.gen.ts
@@ -20,6 +20,11 @@ export const zArbeidOgFrilansRegisterInntektDto = z.object({
     arbeidsgiver: z.string(),
 });
 
+export const zBekreftelseDto = z.object({
+    harGodtattEndringen: z.boolean(),
+    uttalelseFraBruker: z.string().optional(),
+});
+
 export const zDeltakerDto = z.object({
     id: z.string().uuid().optional(),
     deltakerIdent: z.string(),
@@ -27,7 +32,7 @@ export const zDeltakerDto = z.object({
 
 export const zOppgavetype = z.enum(['BEKREFT_ENDRET_PROGRAMPERIODE', 'BEKREFT_AVVIK_REGISTERINNTEKT']);
 
-export const zOppgavetypeDataDto = z.object({});
+export const zOppgavetypeDataDto = z.unknown();
 
 export const zProgramperiodeDto = z.object({
     fomDato: z.string().date(),
@@ -68,6 +73,7 @@ export const zOppgaveDto = z.object({
     oppgaveReferanse: z.string().uuid(),
     oppgavetype: zOppgavetype,
     oppgavetypeData: z.union([zEndretProgramperiodeDataDto, zKontrollerRegisterinntektOppgavetypeDataDto]),
+    bekreftelse: zBekreftelseDto.optional(),
     status: zOppgaveStatus,
     opprettetDato: z.string().datetime(),
     løstDato: z.string().datetime().optional(),
@@ -105,13 +111,13 @@ export const zNavn = z.object({
     etternavn: z.string(),
 });
 
-export const zDeltakerPersonlia = z.object({
+export const zDeltakerPersonalia = z.object({
     id: z.string().uuid().optional(),
     deltakerIdent: z.string(),
     navn: zNavn,
     fødselsdato: z.string().date(),
-    sisteMuligeInnmeldingsdato: z.string().date(),
     førsteMuligeInnmeldingsdato: z.string().date(),
+    sisteMuligeInnmeldingsdato: z.string().date(),
 });
 
 export const zRegisterInntektArbeidOgFrilansDto = z.object({
@@ -147,7 +153,8 @@ export const zEndretProgamperiodeOppgaveDto = z.object({
 });
 
 export const zKontonummerDto = z.object({
-    kontonummer: z.string(),
+    harKontonummer: z.boolean(),
+    kontonummer: z.string().optional(),
 });
 
 export const zRapportPeriodeinfoDto = z.object({
@@ -180,7 +187,7 @@ export const zMeldInnDeltakerResponse = zDeltakelseOpplysningDto;
 
 export const zHentAlleDeltakelserGittDeltakerAktørResponse = zDeltakerOpplysningerDto;
 
-export const zHentDeltakerInfoGittDeltakerResponse = zDeltakerPersonlia;
+export const zHentDeltakerInfoGittDeltakerResponse = zDeltakerPersonalia;
 
 export const zKontrollAvRegisterinntektResponse = zOppgaveDto;
 
@@ -190,7 +197,7 @@ export const zOpprettOppgaveForEndretProgramperiodeResponse = zOppgaveDto;
 
 export const zHentAlleDeltakelserGittDeltakerIdResponse = z.array(zDeltakelseOpplysningDto);
 
-export const zHentDeltakerInfoGittDeltakerIdResponse = zDeltakerPersonlia;
+export const zHentDeltakerInfoGittDeltakerIdResponse = zDeltakerPersonalia;
 
 export const zHentKontonummerResponse = zKontonummerDto;