+++ packages/k9-brukerdialog-prosessering-api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @navikt/k9-brukerdialog-prosessering-api
 
+## 0.0.3
+
+### Patch Changes
+
+- 3e4c469: Pakkeoppdateringer - minor/patch
+
 ## 0.0.2
 
 ### Patch Changes